1.20.4 Support + Adds 1.20.4 support + Refactor Glow class + Rename package + Bump version + Bump NBT API + Revert commit making yml files only generate on debug mode since it breaks new installations.
