+++ build.gradle
@@ -3,8 +3,8 @@ plugins {
     id 'java'
 }
 
-group 'me.yamakaja.commanditems'
-version '2.6.4-pre'
+group 'me.relaxing9.commanditems'
+version '2.6.4'
 
 sourceCompatibility = targetCompatibility = JavaVersion.VERSION_1_8
 
@@ -35,7 +35,7 @@ dependencies {
     implementation "com.fasterxml.jackson.dataformat:jackson-dataformat-yaml:2.15.3"
     implementation "com.fasterxml.jackson.core:jackson-databind:2.15.3"
 
-    implementation "de.tr7zw:item-nbt-api:2.11.3"
+    implementation "de.tr7zw:item-nbt-api:2.12.2"
 }
 
 shadowJar {
