+++ src/main/java/me/relaxing9/commanditems/data/IterationTarget.java
@@ -1,4 +1,4 @@
-package me.yamakaja.commanditems.data;
+package me.relaxing9.commanditems.data;
 
 /**
  * Created by Yamakaja on 27.05.18.
