+++ src/main/java/me/relaxing9/commanditems/data/action/ActionMathExpr.java
@@ -1,4 +1,4 @@
-package me.yamakaja.commanditems.data.action;
+package me.relaxing9.commanditems.data.action;
 
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -9,9 +9,9 @@ import java.util.logging.Level;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-import me.yamakaja.commanditems.CommandItems;
-import me.yamakaja.commanditems.data.ItemDefinition;
-import me.yamakaja.commanditems.interpreter.InterpretationContext;
+import me.relaxing9.commanditems.CommandItems;
+import me.relaxing9.commanditems.data.ItemDefinition;
+import me.relaxing9.commanditems.interpreter.InterpretationContext;
 
 public class ActionMathExpr extends Action {
