+++ src/main/java/me/relaxing9/commanditems/interpreter/ItemExecutor.java
@@ -1,12 +1,12 @@
-package me.yamakaja.commanditems.interpreter;
+package me.relaxing9.commanditems.interpreter;
 
 import java.util.Map;
 
 import org.bukkit.entity.Player;
 
-import me.yamakaja.commanditems.CommandItems;
-import me.yamakaja.commanditems.data.ItemDefinition;
-import me.yamakaja.commanditems.data.action.Action;
+import me.relaxing9.commanditems.CommandItems;
+import me.relaxing9.commanditems.data.ItemDefinition;
+import me.relaxing9.commanditems.data.action.Action;
 
 /**
  * Created by Yamakaja on 26.05.18.
