+++ src/main/java/me/relaxing9/commanditems/parser/ConfigManager.java
@@ -1,4 +1,4 @@
-package me.yamakaja.commanditems.parser;
+package me.relaxing9.commanditems.parser;
 
 import java.io.File;
 import java.io.IOException;
@@ -12,10 +12,10 @@ import org.bukkit.inventory.ItemStack;
 import com.fasterxml.jackson.databind.module.SimpleModule;
 import com.fasterxml.jackson.dataformat.yaml.YAMLMapper;
 
-import me.yamakaja.commanditems.CommandItems;
-import me.yamakaja.commanditems.data.CommandItemsConfig;
-import me.yamakaja.commanditems.data.ItemDefinition;
-import me.yamakaja.commanditems.data.action.Action;
+import me.relaxing9.commanditems.CommandItems;
+import me.relaxing9.commanditems.data.CommandItemsConfig;
+import me.relaxing9.commanditems.data.ItemDefinition;
+import me.relaxing9.commanditems.data.action.Action;
 
 /**
  * Created by Yamakaja on 26.05.18.
